+++ avalanche/benchmarks/scenarios/__init__.py
@@ -1,7 +1,9 @@
 from .generic_scenario import *
-from .dataset_scenario import *
+from avalanche.benchmarks.scenarios.deprecated.dataset_scenario import *
 from avalanche.benchmarks.scenarios.deprecated.classification_scenario import *
 from avalanche.benchmarks.scenarios.deprecated.new_classes import *
 from avalanche.benchmarks.scenarios.deprecated.new_instances import *
+
+from .dataset_scenario import *
 from .exmodel_scenario import *
 from .online_scenario import *
