+++ tests/benchmarks/classic/test_core50.py
