+++ tests/benchmarks/classic/test_tinyimagenet.py
