+++ tests/benchmarks/scenarios/deprecated/test_classification_scenario.py
