+++ tests/benchmarks/scenarios/deprecated/test_custom_streams.py
