+++ tests/benchmarks/scenarios/deprecated/test_generic_scenario.py
