+++ tests/benchmarks/scenarios/deprecated/test_ni_sit_scenario.py
