+++ tests/benchmarks/utils/test_avalanche_classification_dataset.py
