+++ tests/benchmarks/utils/test_avalanche_dataset.py
