+++ tests/training/test_supervised_regression.py
@@ -36,7 +36,7 @@ from avalanche.training.plugins.clock import Clock
 from avalanche.training.plugins import EvaluationPlugin
 from avalanche.training.templates import SupervisedTemplate
 from avalanche.training.utils import trigger_plugins
-from tests.test_dataloaders import get_fast_benchmark
+from benchmarks.utils.test_dataloaders import get_fast_benchmark
 
 logger = logging.getLogger(__name__)
